Split-up tree functionality into separate modules
